Move /searchAll api from '/workflow/searchAll' to '/dashboard/searchAll'
